+++ src/util/index.ts
@@ -2,8 +2,7 @@
 
 The MIT License (MIT)
 
-Copyright (c) 2016 Tom Zoehner
-Copyright (c) 2018 Thomas Bluemel
+Copyright (c) 2020 Tom Zoehner
 
 Permission is hereby granted, free of charge, to any person obtaining a copy
 of this software and associated documentation files (the "Software"), to deal
@@ -25,5 +24,4 @@ SOFTWARE.
 
 */
 
-export { Renderer, IRendererSettings } from "./Renderer";
-export { EMFJSError as Error, loggingEnabled } from "./Helper";
+export { SVG } from "./SVG";