+++ test/utils.js
@@ -1,12 +1,12 @@
 var jsdom = require("jsdom");
-var { JSDOM } = jsdom;
+var {JSDOM} = jsdom;
 var $_$twiz = require('typewiz-core/dist/type-collector-snippet').$_$twiz;
 global.__coverage__ = {};
 
 function stringToArrayBuffer(string) {
     var buffer = new ArrayBuffer(string.length);
     var bufferView = new Uint8Array(buffer);
-    for (var i=0; i<string.length; i++) {
+    for (var i = 0; i < string.length; i++) {
         bufferView[i] = string.charCodeAt(i);
     }
     return buffer;
@@ -18,16 +18,11 @@ function indentHtml(rawHtml) {
         .replace("\n", "");
 }
 
-exports.runRtfjs = function(path, source, callback, errorCallback) {
+exports.runRtfjs = function (path, source, callback, errorCallback) {
     const virtualConsole = new jsdom.VirtualConsole();
     virtualConsole.sendTo(console);
 
-    var dom = new JSDOM(`
-    <script src="../samples/.common/dep/jquery.min.js"></script>
-
-    <script src="../dist/jquery.svg.min.js"></script>
-    <script src="../dist/jquery.svgfilter.min.js"></script>
-
+    new JSDOM(`
     <script src="../dist/WMFJS.bundle.js"></script>
     <script src="../dist/EMFJS.bundle.js"></script>
     <script src="../dist/RTFJS.bundle.js"></script>
@@ -37,8 +32,8 @@ exports.runRtfjs = function(path, source, callback, errorCallback) {
         WMFJS.loggingEnabled(false);
         EMFJS.loggingEnabled(false);
 
-        var baseURL = path + "/";
-        var settings = {
+        const baseURL = path + "/";
+        const settings = {
             onImport: function(relURL, cb) {
                 const file = baseURL + relURL;
                 const ext  = relURL.replace(/^.*\\.([^\\.]+)$/, '$1').toLowerCase();
@@ -54,12 +49,12 @@ exports.runRtfjs = function(path, source, callback, errorCallback) {
                         return cb({error});
                 }
 
-                var request = new XMLHttpRequest();
+                const request = new XMLHttpRequest();
                 request.open("GET", file, true);
                 request.responseType = "arraybuffer";
 
                 request.onload = function (event) {
-                    var blob = request.response;
+                    const blob = request.response;
                     if (blob) {
                         let height = 300;
                         cb({keyword, blob, height});
@@ -71,62 +66,56 @@ exports.runRtfjs = function(path, source, callback, errorCallback) {
 
                 request.send(null);
             }
-        }
+        };
 
         try {
-            var doc = new RTFJS.Document(rtfFile, settings);
+            const doc = new RTFJS.Document(rtfFile, settings);
 
-            var meta = doc.metadata();
+            const meta = doc.metadata();
             doc.render().then(function(htmlElements) {
-                var html = $("<div>").append(htmlElements).html();
+                const div = document.createElement("div");
+                div.append(...htmlElements);
 
-                window.done(meta, html);
+                window.done(meta, div.innerHTML);
             }).catch(error => window.onerror(error))
         } catch (error){
             window.onerror(error)
         }
     </script>
-    `, { resources: "usable",
+    `, {
+        resources: "usable",
         runScripts: "dangerously",
         url: "file://" + __dirname + "/",
         virtualConsole,
         beforeParse(window) {
             window.path = path;
             window.rtfFile = stringToArrayBuffer(source);
-            window.done = function(meta, html){
+            window.done = function (meta, html) {
                 callback(JSON.stringify(meta, null, 4), indentHtml(html), $_$twiz);
             };
             window.onerror = function (error) {
                 errorCallback(error)
             };
-            // Catch exceptions from jquery.svg.min.js
-            window.alert = function (error) {
-                errorCallback(error)
-            };
             window.$_$twiz = $_$twiz;
             window.__coverage__ = global.__coverage__;
-        }});
+        }
+    });
 };
 
-exports.runEmfjs = function(source, callback, errorCallback) {
+exports.runEmfjs = function (source, callback, errorCallback) {
     const virtualConsole = new jsdom.VirtualConsole();
     virtualConsole.sendTo(console);
 
-    var dom = new JSDOM(`
-    <script src="../samples/.common/dep/jquery.min.js"></script>
-
-    <script src="../dist/jquery.svg.min.js"></script>
-    <script src="../dist/jquery.svgfilter.min.js"></script>
-
+    new JSDOM(`
     <script src="../dist/EMFJS.bundle.js"></script>
 
     <script>
         EMFJS.loggingEnabled(false);
 
         try {
-            var width = 600;
-            var height = 400;
-            var settings = {
+            const width = 600;
+            const height = 400;
+            const settings = {
                 width: width + "px",
                 height: height + "px",
                 wExt: width,
@@ -136,54 +125,47 @@ exports.runEmfjs = function(source, callback, errorCallback) {
                 mapMode: 8 // preserve aspect ratio
             };
 
-            var renderer = new EMFJS.Renderer(emfFile);
-            var svg = renderer.render(settings).html();
+            const renderer = new EMFJS.Renderer(emfFile);
+            const svg = renderer.render(settings).innerHTML;
 
             window.done(svg);
         } catch (error){
             window.onerror(error)
         }
     </script>
-    `, { resources: "usable",
+    `, {
+        resources: "usable",
         runScripts: "dangerously",
         url: "file://" + __dirname + "/",
         virtualConsole,
         beforeParse(window) {
             window.emfFile = source;
-            window.done = function(svg){
+            window.done = function (svg) {
                 callback(indentHtml(svg), $_$twiz);
             };
             window.onerror = function (error) {
                 errorCallback(error)
             };
-            // Catch exceptions from jquery.svg.min.js
-            window.alert = function (error) {
-                errorCallback(error)
-            };
             window.$_$twiz = $_$twiz;
             window.__coverage__ = global.__coverage__;
-        }});
+        }
+    });
 };
 
-exports.runWmfjs = function(source, callback, errorCallback) {
+exports.runWmfjs = function (source, callback, errorCallback) {
     const virtualConsole = new jsdom.VirtualConsole();
     virtualConsole.sendTo(console);
 
-    var dom = new JSDOM(`
-    <script src="../samples/.common/dep/jquery.min.js"></script>
-
-    <script src="../dist/jquery.svg.min.js"></script>
-    <script src="../dist/jquery.svgfilter.min.js"></script>
-
+    new JSDOM(`
     <script src="../dist/WMFJS.bundle.js"></script>
 
     <script>
         WMFJS.loggingEnabled(false);
 
         try {
-            var width = 600;
-            var height = 400;
-            var settings = {
+            const width = 600;
+            const height = 400;
+            const settings = {
                 width: width + "px",
                 height: height + "px",
                 xExt: width,
@@ -191,31 +173,29 @@ exports.runWmfjs = function(source, callback, errorCallback) {
                 mapMode: 8 // preserve aspect ratio
             };
 
-            var renderer = new WMFJS.Renderer(wmfFile);
-            var svg = renderer.render(settings).html();
+            const renderer = new WMFJS.Renderer(wmfFile);
+            const svg = renderer.render(settings).innerHTML;
 
             window.done(svg);
         } catch (error){
             window.onerror(error)
         }
     </script>
-    `, { resources: "usable",
+    `, {
+        resources: "usable",
         runScripts: "dangerously",
         url: "file://" + __dirname + "/",
         virtualConsole,
         beforeParse(window) {
             window.wmfFile = source;
-            window.done = function(svg){
+            window.done = function (svg) {
                 callback(indentHtml(svg), $_$twiz);
             };
             window.onerror = function (error) {
                 errorCallback(error)
             };
-            // Catch exceptions from jquery.svg.min.js
-            window.alert = function (error) {
-                errorCallback(error)
-            };
             window.$_$twiz = $_$twiz;
             window.__coverage__ = global.__coverage__;
-        }});
+        }
+    });
 };