+++ webpack/base.config.js
@@ -1,5 +1,4 @@
 const path = require('path');
-const CopyWebpackPlugin = require('copy-webpack-plugin');
 
 module.exports = {
     mode: 'none',
@@ -36,29 +35,8 @@ module.exports = {
             commonjs2: './WMFJS.bundle.js',
             amd: './WMFJS.bundle.js',
             root: 'WMFJS'
-        },
-        jquery: {
-            commonjs: 'jquery',
-            commonjs2: 'jquery',
-            amd: 'jquery',
-            root: '$'
-        },
-        "jquery.svg": {
-            commonjs: './jquery.svg',
-            commonjs2: './jquery.svg',
-            amd: './jquery.svg',
-            root: '$'
-        },
-        "jquery.svgfilter": {
-            commonjs: './jquery.svgfilter',
-            commonjs2: './jquery.svgfilter',
-            amd: './jquery.svgfilter',
-            root: '$'
         }
     },
-    plugins: [
-        new CopyWebpackPlugin([{ from: 'vendor/jquery-svg', context: path.join(__dirname, '..'), force: true}])
-    ],
     stats: {
         modules: false
     }